+++ grpc_stdio.go
@@ -11,11 +11,11 @@ import (
 	"io"
 
 	"github.com/arloliu/go-plugin/internal/plugin"
-	empty "github.com/golang/protobuf/ptypes/empty"
 	hclog "github.com/hashicorp/go-hclog"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
+	"google.golang.org/protobuf/types/known/emptypb"
 )
 
 // grpcStdioBuffer is the buffer size we try to fill when sending a chunk of
@@ -50,7 +50,7 @@ func newGRPCStdioServer(log hclog.Logger, srcOut, srcErr io.Reader) *grpcStdioSe
 
 // StreamStdio streams our stdout/err as the response.
 func (s *grpcStdioServer) StreamStdio(
-	_ *empty.Empty,
+	_ *emptypb.Empty,
 	srv plugin.GRPCStdio_StreamStdioServer,
 ) error {
 	// Share the same data value between runs. Sending this over the wire
@@ -102,7 +102,7 @@ func newGRPCStdioClient(
 	client := plugin.NewGRPCStdioClient(conn)
 
 	// Connect immediately to the endpoint
-	stdioClient, err := client.StreamStdio(ctx, &empty.Empty{})
+	stdioClient, err := client.StreamStdio(ctx, &emptypb.Empty{})
 
 	// If we get an Unavailable or Unimplemented error, this means that the plugin isn't
 	// updated and linking to the latest version of go-plugin that supports